refactor: 장소에 대한 알림 리팩토링 (#308)
* chore: 레거시 코드 제거

* feat: 장소에 대해 알림이 잘 오도록 코드 수정

* refactor: 자신의 위치를 보내는 코드 _app으로 이동

* fix: stroke-width 경고 제거

* refactor: delete magic number in interval function
