--- src/components/index/TodoCalendar/index.tsx
@@ -1,11 +1,0 @@
-import { Container } from './styles';
-
-const TodoCalendar: React.FC = () => {
-  return (
-    <Container>
-      <input type="time" />
-    </Container>
-  );
-};
-
-export default TodoCalendar;
