--- src/components/index/TodoCalendar/styles.tsx
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  width: 20rem;
-  height: 25rem;
-  background-color: #ffffff;
-  border: #eeeeee 0.1rem solid;
-  border-radius: 1rem;
-
-  position: absolute;
-  top: 3rem;
-  right: 0;
-  z-index: 10;
-`;
