--- src/components/index/TodoList/styles.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.main`
-  height: 100%;
-  margin-bottom: 3rem;
-
-  overflow-y: scroll;
-`;
