--- src/components/index/TodoListItem/index.stories.tsx
@@ -1,25 +1,0 @@
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-
-import TodoListItem from '.';
-import { Todo } from '@/shared/types/todo';
-
-export default {
-  title: 'Index/TodoListItem',
-  component: TodoListItem,
-} as ComponentMeta<typeof TodoListItem>;
-
-const todo: Todo = {
-  id: '1',
-  title: 'Todo 1',
-  alarmed: false,
-  done: false,
-};
-
-const Template: ComponentStory<typeof TodoListItem> = (args) => <TodoListItem {...args} />;
-
-export const Basic = Template.bind({});
-
-Basic.args = {
-  id: todo.id,
-  title: todo.title,
-};
