+++ src/shared/types/location.ts
@@ -6,9 +6,3 @@ export interface LocationSearchResult {
   name: string;
   location: [latitude: number, longitude: number];
 }
-
-export interface GeoJson {
-  type: string;
-  name: string;
-  coordinates: Array<number>;
-}
